+++ src/main/java/com/solvd/hospital/models/SickLeave.java
@@ -4,11 +4,11 @@ import java.util.Date;
 
 public class SickLeave {
 
-    private long id;
+    private int id;
 
-    private long medicalCardId;
+    private int medicalCardId;
 
-    private long doctorId;
+    private int doctorId;
 
     private Date from;
 
@@ -22,7 +22,7 @@ public class SickLeave {
 
     }
 
-    public SickLeave (long id, long medicalCardId, long doctorId, Date from, Date to, String diagnosisCode,
+    public SickLeave (int id, int medicalCardId, int doctorId, Date from, Date to, String diagnosisCode,
                       String placeOfDepartment) {
         this.id = id;
         this.medicalCardId = medicalCardId;
@@ -33,27 +33,27 @@ public class SickLeave {
         this.placeOfDepartment = placeOfDepartment;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public long getMedicalCardId() {
+    public int getMedicalCardId() {
         return medicalCardId;
     }
 
-    public void setMedicalCardId(long medicalCardId) {
+    public void setMedicalCardId(int medicalCardId) {
         this.medicalCardId = medicalCardId;
     }
 
-    public long getDoctorId() {
+    public int getDoctorId() {
         return doctorId;
     }
 
-    public void setDoctorId(long doctorId) {
+    public void setDoctorId(int doctorId) {
         this.doctorId = doctorId;
     }
